+++ src/App.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import { WelcomePage } from "./pages/WelcomePage";
 import { WorkshopPreviewPage } from "./pages/WorkshopPage";
+import "./bulma.min.css";
 
 export function App() {
   return (