+++ src/helpers/set-html-overflow.ts
@@ -1,0 +1,3 @@
+export function setHTMLOverflow(overflow: "hidden" | "inherit") {
+  document.getElementsByTagName("html").item(0)!.style.overflow = overflow;
+}