+++ src/components/Markdown.tsx
@@ -1,0 +1,15 @@
+import MarkdownIt from "markdown-it";
+import React from "react";
+const markdownRender = new MarkdownIt();
+
+interface ComponentProps {
+  text: string;
+}
+
+export function Markdown({ text: markdownText }: ComponentProps) {
+  return (
+    <div
+      dangerouslySetInnerHTML={{ __html: markdownRender.render(markdownText) }}
+    ></div>
+  );
+}